chore: Refactor bar layout
Changes the overall layout of the bar
A temporary track title revealer was added to show the track title
temporarily on hover. The `SystemResourcesOrCustomModule` function was
modified to include a `SysRevealer` child that reveals the
`SysResources` box on hover. Finally, the `clientsClassesButton`
function was added to handle displaying the clients classes menu.